SUNLIGHT: apparmor4.0.0 [96/99]: apparmor: fix prompt failing during large down loads
BugLink: https://bugs.launchpad.net/bugs/2028253

Firefox was failing large downloads with a denied message, after a
download was started for two interelated reasons. The first was that
the full set of allowed permissions where not being propogated back
to the object fctx cache, which could result in secondary permission
requests outside of the regular open and create hooks.

Second the atomic_context information was not being correctly propgated
so that some of those secondary checks would not result in a prompt
but a default deny.

Change-Id: I98f9b1f51d6811ca2b1123cec7f57028d25f776f
Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>