SUNLIGHT: apparmor4.0.0 [83/84]: apparmor: fix allow field in notification
The allow field currently is being incorrectly set an incorrect value
that is an intersection of what is allowed and the deny field, This
is effectly the same as setting it to the same value as the deny
field as deny is computed from what is requested and ~allow.

This isn't a problem for prompt as it is only using the deny field,
but it does prevent prompt from having additional context that can
be used to improve its requests or caching.

Fix this by correctly negating the deny field before it is anded with
the request, so that allow and deny do not overlap.

Change-Id: Ic69ed23d433e26d741039771b1dbd3d4d8adafa0
Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>